# Conflicts:
#	documentation/sphinx/source/release-notes/release-notes-700.rst

+++ flow/FastAlloc.cpp
@@ -119,7 +119,7 @@ void setFastAllocatorThreadInitFunction(ThreadInitFunction f) {
 std::atomic<int64_t> g_hugeArenaMemory(0);
 
 double hugeArenaLastLogged = 0;
-std::map<std::string, std::pair<int, int>> hugeArenaTraces;
+std::map<std::string, std::pair<int, int64_t>> hugeArenaTraces;
 
 void hugeArenaSample(int size) {
 	if (TraceEvent::isNetworkThread()) {
